Fix #16
Lambda function stdout is now split on `\n` character to get the last value. This assumes the last value in stdout is the result object from the Lambda function. If it cannot be parsed, it is returned as a string. If stdout has no entries, an empty string will be returned to the caller.