allow "invoked function arn" to be specified from env vars
This is to support the use case of invoking lambda using a specific
alias. In this case, the arn is expected to contain the alias, while
the function version should still point to the actual version.

Currently the invoked arn is always derived from the function name +
version, so it's not possible to support the above use case.

Contributed by @quantcast