Support passing extra arguments to bootstrap file of 'provided' runtimes
AWS SAM CLI will use this mechanism to customize what and how the bootstrap gets invoked.
We want to use this to start the language runtime or bootstrap in debug mode.

For Python 3.7, we will invoke the entrypoint as follows:
```bash
/var/rapid/init \
    --bootstrap /var/lang/bin/python3.7 \
    --bootstrap-args '["-u", "/tmp/lambci_debug_files/pydevd.py", "--multiprocess", "--port", "58381", "--file", "/var/runtime/bootstrap"]'
```

For other programming languages (say Nodejs), we might need to create a separate bootstrap
file that starts the node interpretor with debug args. In this case I anticipate us to invoke
using a command similar to this:

```bash
/var/rapid/init \
    --bootstrap /var/runtime/bootstrap.debug \
    --bootstrap-args '["--inspect-brk", "0.0.0.0:9090"]'
```